% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_id.R
\name{audio_id}
\alias{audio_id}
\title{Create an audio id by row at data frame.}
\usage{
audio_id(x, i = 5, drop_fn = TRUE)
}
\arguments{
\item{x}{A data frame.}
}
\description{
Create an audio id by row at data frame.
}
