% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_time.R
\name{audio_time}
\alias{audio_time}
\title{Returns the total time of audio files in seconds}
\usage{
audio_time(x, get.id = FALSE, recursive = FALSE, i = 4)
}
\arguments{
\item{x}{Either a WAV file or a directory containing WAV files.}

\item{get.id}{Logical. Should the ID must be extracted from file name? Default: \code{FALSE}.}

\item{recursive}{Logical.}

\item{i}{ID position in file name. Default: \code{4}.}
}
\description{
Returns the total time of audio files in seconds
}
