% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_svm.R
\name{class_svm}
\alias{class_svm}
\title{Fits and forecast SVM models, serial and parallelized.}
\usage{
class_svm(df, modelo, filtro = c("id"), percTreino = 0.5,
  setSeed = 1, custo = 1, gama = 1, paralelo = T, simbolico = F,
  restart = F)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{modelo}{Character containing the model structure.}

\item{filtro}{Column(s) to filter the samples.}

\item{percTreino}{Percentage of the database used to train the model, filtered by \code{filtro}.}

\item{setSeed}{Specified seed for the pseudo-random parts.}

\item{custo}{Cost of constraints violation (default: 1)—it is the ‘C’-constant of the regularization term in the Lagrange formulation.}

\item{gama}{Needed for all kernels except linear (default: 1/(data dimension))}

\item{paralelo}{Logical. Should the process be parallelized?}

\item{simbolico}{Logical. Should the symbolic confusion matrix be printed?}

\item{restart}{Logical. Should the R session be restarted? (It frees memory)}
}
\value{
\code{$fcast} predicted time series using the model that minimizes the forecasting mean square error.
\code{$runtime} running time.
\code{mse.pred} mean squared error of prediction. Used to decide the best model.
}
\description{
Fits and forecast SVM models, serial and parallelized.
}
\examples{
library(voice)
}
\references{
Vapnik, Vladimir (2000). The Nature of Statistical Learning Theory, Springer.
}
