% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_audio.R
\name{cut_audio}
\alias{cut_audio}
\title{Cut audio vectors}
\usage{
cut_audio(x, byvar = x)
}
\arguments{
\item{x}{A vector containing the feature to be cut by \code{byvar}.}

\item{byvar}{A vector containing the variable to cut by.}
}
\description{
Cut vectors
}
\examples{
library(voice)
# get path to audio file
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
pattern = glob2rx('*.wav'), full.names = TRUE)
# Media dataset
M <- extract_features(path2wav)
cut_audio(M$f0)
cut_audio(M$gain, M$f0)
}
