% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarization.R
\name{diarization}
\alias{diarization}
\title{Extract features from WAV audios using Python's Parselmouth library.}
\usage{
diarization(directory)
}
\arguments{
\item{directory}{A directory/folder containing WAV files.}
}
\description{
Extract features from WAV audios using Python's Parselmouth library.
}
