% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_py.R
\docType{data}
\name{directory}
\alias{directory}
\title{Extract features from WAV audios using Python's Parselmouth library.}
\format{
An object of class \code{character} of length 1.
}
\usage{
directory
}
\arguments{
\item{\code{directory}}{A directory/folder containing WAV files.}

\item{\code{filesRange}}{The desired range of directory files (default: \code{NULL}, i.e., all files).}

\item{\code{features}}{Vector of features to be extracted. (default: 'f0' (pitch),'formants' (F1:F8)).}
}
\value{
\code{char} vector containing the expanded models.
}
\description{
Extract features from WAV audios using Python's Parselmouth library.
}
\details{
The function uses the \code{getwd()} folder to write the temp files.
}
\examples{
library(voice)
path2wav <- list.files(system.file("extdata", package = "wrassp"),
pattern <- glob2rx("*.wav"), full.names = TRUE)
efp <- extract_features_py(dirname(path2wav))
efp
table(efp$file_name)

# limiting filesRange
efpl <- extract_features_py(dirname(path2wav), filesRange = 3:6)
efpl
table(efpl$file_name)

}
\keyword{datasets}
