% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration}
\alias{duration}
\title{Gives the duration of sequences in number of lines and miliseconds.}
\usage{
duration(x, windowShift = 5)
}
\arguments{
\item{x}{A vector containing symbols and \code{NA}.}

\item{windowShift}{\code{= <dur>} set analysis window shift to <dur>ation in ms (default: 5.0).}
}
\value{
A tibble with frequencies for equal-tempered scale, A4 = 440 Hz.
}
\description{
Gives the duration of sequences in number of lines and miliseconds.
}
\examples{
library(voice)
duration(letters)
duration(c('a','a','a',letters,'z'))

nts <- c('NA','NA','A3','A3','A3','A3','A#3','B3','B3','C4','C4','C4','C4','
C4','C4','C#4','C4','C4','C4','B3','A#3','NA','NA','NA','NA','NA','NA','NA',
'NA','NA','NA','NA','NA','NA','NA','NA','NA','NA','NA','D4','D4','D4','C#4',
'C#4','C#4','C4','C4','B3','B3','A#3','A#3','A3','A3','G3','G#3','G3','F#3',
'F#3','G3','NA','F#3','F#3','F#3','F#3','F#3','G3','F#3','F#3','F#3','F#3',
'F3','F3','F3','F3','E3','E3','E3','D#3','D#3','D#3','D#3','D#3','D3','D3',
'D#3','D#3','D#3','D#3','D#3','D3','D#3','D3','D3','D3','D#3','D3','D#3','D#3',
'D#3','D#3')
duration(nts)
}
