% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_summary.R
\name{feat_summary}
\alias{feat_summary}
\title{Returns summary measures of voice::extract_features}
\usage{
feat_summary(
  x,
  filesRange = NULL,
  features = "f0",
  extraFeatures = FALSE,
  gender = "u",
  windowShift = 5,
  numFormants = 8,
  numcep = 12,
  dcttype = c("t2", "t1", "t3", "t4"),
  fbtype = c("mel", "htkmel", "fcmel", "bark"),
  resolution = 40,
  usecmp = FALSE,
  mc.cores = 1,
  full.names = TRUE,
  recursive = FALSE,
  check.mono = TRUE,
  stereo2mono = TRUE,
  overwrite = FALSE,
  freq = 44100,
  round.to = 4,
  get.id = FALSE,
  i = 4
)
}
\arguments{
\item{x}{Either a WAV file or a directory containing WAV files.}

\item{get.id}{Logical. Should the ID must be extracted from file name? Default: \code{FALSE}.}

\item{i}{ID position in file name. Default: \code{4}.}
}
\description{
Returns summary measures of voice::extract_features
}
