% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samp.rate.R
\name{get_samp.rate}
\alias{get_samp.rate}
\title{Get sample rate}
\usage{
get_samp.rate(x)
}
\arguments{
\item{fromWav}{A directory/folder containing WAV files.}
}
\description{
Get sample rate from WAV file.
}
