% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_audio.R
\name{has_audio}
\alias{has_audio}
\title{Indicates if an ID has associated audio files}
\usage{
has_audio(x, mediaDir, id = NULL, get.id = FALSE, i = 4, to.int = TRUE)
}
\arguments{
\item{x}{A data frame to be tagged with media information.}

\item{mediaDir}{Directory containing media files. Currently accepts audio files with WAV and MP3 formats.}

\item{id}{Column containing the ID to be searched in mediaDir file names. Default: \coe{NULL}, i.e., uses the first column.}

\item{i}{ID position in \code{col}. Default: \code{4}.}
}
\description{
Indicates if an ID has associated audio files
}
