% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_file.R
\name{id_file}
\alias{id_file}
\title{Create an audio id by row at data frame.}
\usage{
id_file(x, col = NULL, pattern = "[_.]", i = 5, drop_col = FALSE)
}
\arguments{
\item{x}{A data frame.}

\item{col}{A column containing the ID to be extracted. If \code{NULL}, the first column is used. Default: \code{NULL}.}

\item{pattern}{Pattern to split \code{col}. Default: \code{'[_.]'}.}

\item{i}{ID position in \code{col}. Default: \code{5}.}

\item{drop_col}{Logical. Should \code{col} be removed? Default: \code{FALSE}.}
}
\description{
Create an audio id by row at data frame.
}
