% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_path.R
\docType{data}
\name{id_path}
\alias{id_path}
\title{Sample IDs and paths}
\format{
A tibble with 34,425 rows and 2 variables:
\describe{
  \item{client_id}{Hash containing client ID.}
  \item{path}{Media file names.}
}
}
\usage{
\code{data(id_path)}
}
\description{
A dataset containing sample IDs and paths used in
Zabala (2022) voice: new approaches to audio analysis.
The considered sample contains 34,425 rows associated with
838 IDs ($p_{s} \approx 2.4\%$).
}
\keyword{datasets}
