% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.audio}
\alias{is.audio}
\title{Match string for audio suffix}
\usage{
is.audio(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Given a character vector, returns a logical vector indicating
which elements have a valid audio file extension.
}
