% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.hosted}
\alias{is.hosted}
\title{Return strings with a URL scheme}
\usage{
is.hosted(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Given a character vector, returns a logical indicating whether the
URLs in the vector respond without error.
}
