% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.local}
\alias{is.local}
\title{Return strings without a URL scheme}
\usage{
is.local(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Given a character vector, returns a logical indicating whether the
paths in the vector point to existing local files.
}
