% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.url}
\alias{is.url}
\title{Match string for URL prefix}
\usage{
is.url(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Given a character vector, returns a logical vector indicating
which elements have a URL scheme.
}
