% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.video}
\alias{is.video}
\title{Match string for video suffix}
\usage{
is.video(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Given a character vector, returns a logical vector indicating
which elements have a valid video file extension.
}
