% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_can.R
\name{is_can}
\alias{is_can}
\title{Is a database of the canonical type?}
\usage{
is_can(x, subj.id = subj.id)
}
\arguments{
\item{x}{A data frame to be evaluated.}

\item{subj.id}{Column containing the subject ID. Default: \code{NULL}, i.e., uses the first column.}
}
\description{
Is a database of the canonical type?
}
