% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ext.R
\name{is_ext}
\alias{is_ext}
\title{Is a database of the extended type?}
\usage{
is_ext(x, subj.id = NULL)
}
\arguments{
\item{x}{A data frame to be evaluated.}

\item{subj.id}{Column containing the subject ID. Default: \coe{NULL}, i.e., uses the first column.}
}
\description{
Is a database of the extended type?
}
