% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_files.R
\docType{data}
\name{media-files}
\alias{media-files}
\alias{mp3}
\alias{mp4}
\alias{png}
\title{These functions are sourced from the `embedr` package by Michael McCarthy, under
MIT License: https://github.com/mccarthy-m-g/embedr/blob/master/LICENSE.md
This inclusion is temporary and will be discontinued once `embedr` is available on CRAN.
See https://github.com/mccarthy-m-g/embedr for more details.}
\format{
An object of class \code{character} of length 1.
}
\usage{
mp3
}
\description{
Example Media Files
}
\details{
Example media files included with embedr.

 - `mp3`: MP3 audio
 - `mp4`: MP4 video
 - `png`: PNG thumbnail
}
\keyword{datasets}
