% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_filter.R
\name{na_filter}
\alias{na_filter}
\title{NA filter}
\usage{
na_filter(x, max.loss = 1)
}
\arguments{
\item{x}{An object of class \code{aggr} from \code{VIM} package.}

\item{max.loss}{A directory/folder containing RTTM files.}
}
\description{
Filters
}
\examples{
\dontrun{
library(voice)
# http://archive.ics.uci.edu/ml/datasets/Arrhythmia
dat <- 'http://archive.ics.uci.edu/ml/machine-learning-databases/00579/MI.data'
bd <- read.table(dat, sep = ',')
bd[bd=='?'] <- NA
}
}
