% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_freq.R
\name{notes_freq}
\alias{notes_freq}
\title{Returns a tibble with frequencies for equal-tempered scale, A4 = 440 Hz.}
\usage{
notes_freq()
}
\description{
Returns a tibble with frequencies for equal-tempered scale, A4 = 440 Hz.
}
\details{
The symbol 'ß' is being used to represent a sharp note, the higher in pitch by one semitone. The symbol 'b' represents a flat note, the lower in pitch by one semitone.
}
\examples{
library(voice)
notes_freq()
}
\references{
https://pages.mtu.edu/~suits/notefreqs.html
}
