% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_summary.R
\name{notes_summary}
\alias{notes_summary}
\title{Returns summary measures of the spoken notes in semitones}
\usage{
notes_summary(x, get.id = FALSE, i = 4, recursive = FALSE)
}
\arguments{
\item{x}{Either a WAV file or a directory containing WAV files.}

\item{get.id}{Logical. Should the ID must be extracted from file name? Default: \code{FALSE}.}

\item{i}{ID position in file name. Default: \code{4}.}

\item{recursive}{Logical. Should the listing recursively into directories? (default: \code{FALSE}) Used by \code{base::list.files}.}
}
\description{
Returns summary measures of the spoken notes in semitones
}
