% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octave.R
\name{octave}
\alias{octave}
\title{Convert SPN to standard octave}
\usage{
octave(x, to_lower = FALSE, spacing = TRUE)
}
\arguments{
\item{x}{A vector containing a note in SPN (Scientific Pitch Notation).}

\item{to_lower}{Logical. Should the string be lower case? Default: \code{FALSE}.}

\item{spacing}{Logical. Should the strin return spaces between notes? Default: \code{TRUE}.}
}
\description{
Convert SPN to standard octave.
}
\examples{
library(voice)
octave('C4')
octave('C5')
all.equal(octave('C4'), octave('C4'))
octave('C4', to_lower = TRUE)
octave(c('C4','D#7','E2'))
octave(c('C4','D#7','E2'), to_lower = TRUE)
octave(c('C4','D#7','E2'), spacing = FALSE)
octave(c('C4','D#7','E2'), to_lower = TRUE, spacing = FALSE)
}
