% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_q.R
\name{plot_q}
\alias{plot_q}
\title{Plot quality measures.}
\usage{
plot_q(
  fit,
  title = NULL,
  ymin = NULL,
  ymax = NULL,
  measure = c("Accuracy", "Kappa", "Sensitivity", "Specificity", "Pos.Pred.Value",
    "Neg.Pred.Value", "Precision", "Recall", "F1", "Prevalence", "Detection.Rate",
    "Detection.Prevalence", "Balanced.Accuracy")
)
}
\arguments{
\item{fit}{A data frame containing quality measures in columns.}

\item{title}{Desired plot title.}

\item{ymin, ymax}{Y label range.}

\item{measure}{Desired measures.}
}
\description{
Get bit from WAV file.
}
