% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rttm.R
\name{read_rttm}
\alias{read_rttm}
\title{Read RTTM files}
\usage{
read_rttm(fromRttm)
}
\arguments{
\item{fromRttm}{A directory/folder containing RTTM files.}
}
\description{
Read Rich Transcription Time Marked (RTTM) files in \code{fromRttm} directory.
}
\details{
The Rich Transcription Time Marked (RTTM) files are space-delimited text files containing one turn per line defined National Institute of Standards and Technology. Each line containing ten fields: '\n'
\code{type} Type: segment type; should always by SPEAKER. '\n'
\code{file} File ID: file name; basename of the recording minus extension (e.g., rec1_a). '\n'
\code{chnl} Channel ID: channel (1-indexed) that turn is on; should always be 1. '\n'
\code{tbeg} Turn Onset -- onset of turn in seconds from beginning of recording. '\n'
\code{tdur} Turn Duration -- duration of turn in seconds.'\n'
\code{ortho} Orthography Field -- should always by <NA>.'\n'
\code{stype} Speaker Type -- should always be <NA>.'\n'
\code{name} Speaker Name -- name of speaker of turn; should be unique within scope of each file.'\n'
\code{conf} Confidence Score -- system confidence (probability) that information is correct; should always be <NA>.'\n'
\code{slat} Signal Lookahead Time -- should always be <NA>.'\n'
}
\examples{
library(voice)
download.file...
}
\references{
https://www.nist.gov/system/files/documents/itl/iad/mig/KWS15-evalplan-v05.pdf
}
\seealso{
\code{voice::enrich_rttm}
}
