% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm0.R
\name{rm0}
\alias{rm0}
\title{Transforms \code{n} sets of \code{m>n} zeros (alternated with sets of non zeros) into \code{n} sets of \code{n} zeros.}
\usage{
rm0(x)
}
\arguments{
\item{x}{A vector or time series.}
}
\value{
Vector with n zeros.
}
\description{
Transforms \code{n} sets of \code{m>n} zeros (alternated with sets of non zeros) into \code{n} sets of \code{n} zeros.
}
\examples{
(v0 <- c(1:20,rep(0,10)))
rm0(v0)

(v0 <- c(rep(0,10),1:20))
rm0(v0)

(v0 <- c(1:5,rep(0,10),1:10,rep(0,5),10:20,rep(0,10)))
rm0(v0)

(v0 <- rep(0,10))
rm0(v0)

(v0 <- c(0:10))
rm0(v0)

(v0 <- c(rep(0,10),1:10,rep(0,5),10:20,rep(0,10)))
rm0(v0)
}
