% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowProp.R
\name{rowProp}
\alias{rowProp}
\title{Form Row Proportions.}
\usage{
rowProp(x, pts = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A matrix, data frame or tibble.}

\item{pts}{Maximum number of points by question. Default \code{NULL}, i.e., automatic.}

\item{na.rm}{Logical. Must the NAs be removed? Default \code{TRUE}.}
}
\description{
Form Row Proportions.
}
\examples{
library(voice)
nr <- 50
nc <- 10
x <- matrix(sample(c(NA,0:3), nr*nc, replace = TRUE), nrow = nr)
colnames(x) <- paste0('V', 1:nc)
rowProp(x)
rowProp(x, na.rm = F)
}
