% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.R
\name{rp}
\alias{rp}
\title{Prenda's rule. Returns \code{2t+p}.}
\usage{
rp(true, pred)
}
\arguments{
\item{true}{A number.}

\item{pred}{A number.}
}
\value{
\code{t,p -> 2t+p}

\code{0,0 -> 0}

\code{0,1 -> 1}

\code{1,0 -> 2}

\code{1,1 -> 3}

\code{10,12 -> 32}

\code{12,10 -> 34}
}
\description{
Prenda's rule. Returns \code{2t+p}.
}
\examples{
rp(0,0) # 0
rp(0,1) # 1
rp(1,0) # 2
rp(1,1) # 3
rp(10,12) # 32
rp(12,10) # 34
}
