% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitw.R
\name{splitw}
\alias{splitw}
\title{Split Wave}
\usage{
splitw(
  fromWav,
  fromRttm = NULL,
  toSplit = NULL,
  autoDir = TRUE,
  output = "wave",
  filesRange = NULL,
  full.names = TRUE,
  recursive = FALSE,
  min.time = 0.4
)
}
\arguments{
\item{fromWav}{A directory/folder containing WAV files.}

\item{fromRttm}{A directory/folder containing RTTM files. Default: \code{NULL}.}

\item{toSplit}{A directory/folder to write generated files. Default: \code{NULL}.}

\item{autoDir}{Logical. Must the directories tree must be created? Default: \code{TRUE}. See 'Details'.}

\item{output}{character string, the class of the object to return, either 'wave' or 'list'.}

\item{filesRange}{The desired range of directory files (default: \code{NULL}, i.e., all files).}

\item{full.names}{Logical. If \code{TRUE}, the directory path is prepended to the file names to give a relative file path. If \code{FALSE}, the file names (rather than paths) are returned. (default: \code{TRUE}) Used by \code{base::list.files}.}

\item{recursive}{Logical. Should the listing recursively into directories? (default: \code{FALSE}) Used by \code{base::list.files}.}

\item{min.time}{Minimum slice length in seconds. (default: \code{0.4})}
}
\description{
Split WAV files in \code{fromWav} folder using (same names) RTTM files as guidance.
}
\details{
When \code{autoDir = TRUE}, the following directories are created: \code{'../mp3'},\code{'../rttm'}, \code{'../split'} and \code{'../musicxml'}. Use \code{getwd()} to find the parent directory \code{'../'}.
}
\examples{
library(voice)
wavDir <- list.files(system.file('extdata', package = 'wrassp'),
                     pattern <- glob2rx('*.wav'), full.names = TRUE)
poetry(dirname(path2wav), pycall = '/home/linuxbrew/.linuxbrew/bin/python3.9') # Linux
splitw(dirname(path2wav))
}
