% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe audio to text}
\usage{
transcribe(
  file,
  server = "http://localhost:8080",
  wait = TRUE,
  poll_interval = 2,
  timeout = 300
)
}
\arguments{
\item{file}{Audio file to trinscribe.}

\item{server}{Default: http://localhost:8080.}

\item{wait}{Default: \code{TRUE}.}

\item{poll_interval}{Default: 2.}

\item{timeout}{Default: 300.}
}
\description{
Transcribe audio to text.
}
\references{
https://github.com/rishikanthc/scriberr
}
