% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url.exists}
\alias{url.exists}
\title{These functions are sourced from the `embedr` package by Michael McCarthy, under
MIT License: https://github.com/mccarthy-m-g/embedr/blob/master/LICENSE.md
This inclusion is temporary and will be discontinued once `embedr` is available on CRAN.
See https://github.com/mccarthy-m-g/embedr for more details.}
\usage{
url.exists(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
`TRUE` if the URL responds without error, otherwise `FALSE`.
}
\description{
Check if URL exists
}
\details{
Given a character string, returns a logical vector indicating
whether a request for a specific URL responds without error.
}
