% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_list.R
\name{write_list}
\alias{write_list}
\title{Writes a list to a path.}
\usage{
write_list(x, directory)
}
\arguments{
\item{x}{A list.}

\item{directory}{A directory.}
}
\description{
Writes a list to a path.
}
