% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_time.R
\name{audio_time}
\alias{audio_time}
\title{Returns the total time of audio files in seconds}
\usage{
audio_time(x, filesRange = NULL, recursive = FALSE)
}
\arguments{
\item{x}{Either a WAV file or a directory containing WAV files.}

\item{filesRange}{The desired range of directory files (default: \code{NULL}, i.e., all files).}

\item{recursive}{Logical.}
}
\description{
Returns the total time of audio files in seconds
}
