% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_svm.R
\name{class_svm}
\alias{class_svm}
\title{Fits, forecasts and gets performance from SVM models, serial and parallelized.}
\usage{
class_svm(
  x,
  model,
  filtro = c("id"),
  percTreino = 0.5,
  setSeed = 1,
  custo = 1,
  gama = 1,
  paralelo = TRUE,
  print.cm = TRUE
)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{model}{Character containing the model structure. See \code{expand_model}.}

\item{filtro}{Column(s) to filter the samples.}

\item{percTreino}{Percentage of the database used to train the model, filtered by \code{filtro}.}

\item{setSeed}{Specified seed for the pseudo-random parts.}

\item{custo}{Cost of constraints violation (default: 1)—it is the ‘C’-constant of the regularization term in the Lagrange formulation.}

\item{gama}{Needed for all kernels except linear (default: 1/(data dimension))}

\item{paralelo}{Logical. Should the process be parallelized?}

\item{print.cm}{Logical. Should the confusion matrix be shown?}
}
\value{
\code{fcast} predicted time series using the model that minimizes the forecasting mean square error.

\code{runtime} running time.

\code{mse.pred} mean squared error of prediction. Used to decide the best model.
}
\description{
Fits, forecasts and gets performance from SVM models, serial and parallelized.
}
\examples{
library(voice)
}
\references{
Vapnik, Vladimir (2000). The Nature of Statistical Learning Theory, Springer.
}
