% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{conv}
\alias{conv}
\title{Convolute vectors.}
\usage{
conv(x, compact.to = 0.1, drop.zeros = T)
}
\arguments{
\item{x}{A vector or time series.}

\item{drop.zeros}{Logical. Drop repeated zeros or compress to 1 zero per null set?}

\item{compact}{Percentage of remaining points after compactation. If equals to 1 and keep.zeros = T, the original vector is presented.}
}
\value{
A vector of convoluted values with length near to \code{compact.to*length(x)}.
}
\description{
Convolute vectors.
}
\examples{
conv(1:100, compact.to = 0.2)
(v0 <- c(1:5,rep(0,10),1:10,rep(0,5),10:20,rep(0,10)))
conv(v0, 0.1)
conv(v0, 0.2)
conv(v0, 0.5)
conv(v0, drop.zeros = F)

(v0 <- c(rep(0,10),1:10,rep(0,5),10:20,rep(0,10)))
conv(v0)

(v0 <- c(rep(0,10),1:20, rep(0,4)))
conv(v0)
plot(v0); points(as.numeric(names(conv(v0))), conv(v0), type = 'l')

(v0 <- c(1:10,rep(0,99),1:66, rep(0,2)))
conv(v0)
conv(v0, 1, T)
}
\seealso{
\code{rm0}
}
