% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_df.R
\name{conv_df}
\alias{conv_df}
\title{Convolute data frames.}
\usage{
conv_df(
  x,
  compact.to,
  colnum = NULL,
  id = "id",
  by.filter = id,
  drop.x = FALSE,
  drop.zeros = FALSE,
  to.data.frame = TRUE,
  round.off = NULL,
  mc.cores = parallel::detectCores()
)
}
\arguments{
\item{x}{A data frame.}

\item{compact.to}{Percentage of remaining points after compactation. If equals to 1 and keep.zeros = T, the original vector is presented.}

\item{colnum}{A \code{char} vector indicating the numeric colnames. If \code{NULL}, uses the columns of the \code{numeric} class.}

\item{id}{The identification column. Default: \code{'id'}.}

\item{by.filter}{A \code{char} indicating the column to filter by. If \code{NULL} uses the \code{id} content.}

\item{drop.x}{Logical. Drop columns containing .x? Default: \code{'FALSE'}.}

\item{drop.zeros}{Logical. Drop repeated zeros or compress to 1 zero per null set? Default: \code{'FALSE'}.}

\item{to.data.frame}{Logical. Should the return be a data frame? If \code{F} returns a list. Default: \code{'TRUE'}.}

\item{round.off}{Number of decimal places of the convoluted vector. Default: \code{'NULL'}.}

\item{mc.cores}{The number of cores to mclapply. By default uses \code{parallel::detectCores()}.}
}
\value{
A vector of convoluted values with length near to \code{compact.to*length(x)}.
}
\description{
Convolute data frames.
}
\examples{
library(voice)

# get path to audio file
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
pattern <- glob2rx('*.wav'), full.names = TRUE)

# getting all the 1092 features
ef <- extract_features(dirname(path2wav), features = c('f0','formants',
'zcr','mhs','rms','gain','rfc','ac','cep','dft','css','lps','mfcc'),
mc.cores = 1)

(cef.df <- conv_df(ef, 0.1, id = 'file_name', mc.cores = 1))
(cef.df2 <- conv_df(ef, 0.1, id = 'file_name', drop.x = TRUE, mc.cores = 1))

dim(ef)
dim(cef.df)
dim(cef.df2)
(cef.list <- conv_df(ef, 0.1, id = 'file_name', to.data.frame = FALSE, mc.cores = 1))
}
\seealso{
\code{conv}, \code{conv_mc}
}
