% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_df.R
\name{conv_df}
\alias{conv_df}
\title{Convolute data frames.}
\usage{
conv_df(
  x,
  compact.to,
  colnum = NULL,
  id = "id",
  by.filter = id,
  drop.x = F,
  drop.zeros = F,
  to.data.frame = T,
  round.off = NULL,
  mc.cores = parallel::detectCores()
)
}
\arguments{
\item{\code{x}}{A data frame.}

\item{\code{compact.to}}{Percentage of remaining points after compactation. If equals to 1 and keep.zeros = T, the original vector is presented.}

\item{\code{colnum}}{A \code{char} vector indicating the numeric colnames. If \code{NULL}, uses the columns of the \code{numeric} class.}

\item{\code{id}}{The identification column. Default: \code{'id'}.}

\item{\code{by.filter}}{A \code{char} indicating the column to filter by. If \code{NULL} uses the \code{id} content.}

\item{\code{drop.x}}{Logical. Drop columns containing .x? Default: \code{'FALSE'}.}

\item{\code{drop.zeros}}{Logical. Drop repeated zeros or compress to 1 zero per null set? Default: \code{'FALSE'}.}

\item{\code{to.data.frame}}{Logical. Should the return be a data frame? If \code{F} returns a list. Default: \code{'TRUE'}.}

\item{\code{round.off}}{Number of decimal places of the convoluted vector. Default: \code{'NULL'}.}

\item{\code{mc.cores}}{The number of cores to mclapply. By default uses \code{parallel::detectCores()}.}
}
\value{
A vector of convoluted values with length near to \code{compact.to*length(x)}.
}
\description{
Convolute data frames.
}
\examples{
library(voice)

x <- dat \%>\%
mutate_each(as.factor, id:anyep_diff_w1)
(cx.df <- conv_df(x, 0.1))
(cx.df2 <- conv_df(x, 0.1, drop.x = T))
dim(x)
dim(cx.df)
dim(cx.df2)
(cx.list <- conv_df(x, 0.1, to.data.frame = F))
}
\seealso{
\code{conv}, \code{conv_mc}
}
