% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_mc.R
\name{conv_mc}
\alias{conv_mc}
\title{Convolute vectors multicore.}
\usage{
conv_mc(x, columns, compact.to, drop.zeros, mc.cores = detectCores())
}
\arguments{
\item{\code{x}}{A vector or time series.}

\item{\code{columns}}{A \code{char} vector indicating the colnames. If \code{NULL}, uses the columns of the \code{numeric} class.}

\item{\code{by.filter}}{A \code{char} indicating the column to filter by. If \code{NULL} the filter is off.}

\item{\code{compact.to}}{Percentage of remaining points after compactation. If equals to 1 and keep.zeros = T, the original vector is presented.}

\item{\code{drop.zeros}}{Logical. Drop repeated zeros or compress to 1 zero per null set?}

\item{\code{to.data.frame}}{Logical. Should the return be a data frame? If \code{F} returns a list.}

\item{\code{mc.cores}}{The number of cores to mclapply.}
}
\value{
A vector of convoluted values with length near to \code{compact.to*length(x)}.
}
\description{
Convolute vectors multicore.
}
\examples{
conv_mc(1:100, compact.to = 0.2)
(v0 <- c(1:5,rep(0,10),1:10,rep(0,5),10:20,rep(0,10)))
conv(v0, 0.1)
conv(v0, 0.2)
conv(v0, 0.5)
conv(v0, drop.zeros = F)
}
\seealso{
\code{conv} \code{conv_df}
}
