% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarization.R
\name{diarization}
\alias{diarization}
\title{Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.}
\usage{
diarization(directory)
}
\arguments{
\item{directory}{A directory/folder containing WAV files.}
}
\description{
Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.
}
