% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarization.R
\name{diarization}
\alias{diarization}
\title{Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.}
\usage{
diarization(
  from,
  to = paste0(from, "/rttm/"),
  pycall = "~/miniconda3/envs/py38phdz/bin/python"
)
}
\arguments{
\item{from}{A directory/folder containing WAV files.}

\item{to}{A directory/folder to write RTTM files.}

\item{pycall}{Python call.}
}
\description{
Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.
}
\examples{
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
pattern <- glob2rx('*.wav'), full.names = TRUE)
dir.create(paste0(dirname(path2wav)[1], '/rttm/'))
diarization(dirname(path2wav)[1])
}
