% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarization.R
\name{diarization}
\alias{diarization}
\title{Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.}
\usage{
diarization(
  from,
  to = paste0(from, "/rttm/"),
  pycall = "~/miniconda3/envs/py38phdz/bin/python"
)
}
\arguments{
\item{directory}{A directory/folder containing WAV files.}

\item{extension}{The extension of audio files to be extracted. Case sensitive.}
}
\description{
Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.
}
\examples{
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
pattern <- glob2rx('*.wav'), full.names = TRUE)
diarization(path2wav, '.wav')
}
