% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diarization.R
\name{diarization}
\alias{diarization}
\title{Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.}
\usage{
diarization(
  directory,
  extension,
  condacall = "~/miniconda3/envs/py38phdz/bin/python"
)
}
\arguments{
\item{directory}{A directory/folder containing WAV files.}

\item{extension}{The extension of audio files to be extracted. Case sensitive.}
}
\description{
Diarization (who speaks when?) from WAV audios using Python's pyannote-audio library.
}
