% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_py.R
\name{extract_features_py}
\alias{extract_features_py}
\title{Extract features from WAV audios using Python's Parselmouth library.}
\usage{
extract_features_py(
  directory,
  features = c("f0", "formants"),
  full.names = TRUE,
  recursive = FALSE
)
}
\arguments{
\item{\code{directory}}{A directory/folder containing WAV files.}

\item{\code{features}}{Vector of features to be extracted. (default: 'f0','formants','zcr','mhs','rms','gain','rfc','ac','mfcc').}
}
\value{
\code{char} vector containing the expanded models.
}
\description{
Extract features from WAV audios using Python's Parselmouth library.
}
\details{
The function uses the \code{getwd()} folder to write the temp files.
}
\examples{
path2wav <- list.files(system.file("extdata", package = "wrassp"),
pattern <- glob2rx("*.wav"), full.names = TRUE)
extract_features_py(dirname(path2wav)[1])
}
