% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_summary.R
\name{feat_summary}
\alias{feat_summary}
\title{Returns summary measures of voice::extract_features}
\usage{
feat_summary(
  x,
  groupBy = "wav_path",
  wavPath = unique(x$wav_path),
  filesRange = NULL,
  features = "f0",
  gender = "u",
  windowShift = 5,
  numFormants = 8,
  numcep = 12,
  dcttype = c("t2", "t1", "t3", "t4"),
  fbtype = c("mel", "htkmel", "fcmel", "bark"),
  resolution = 40,
  usecmp = FALSE,
  mc.cores = 1,
  full.names = TRUE,
  recursive = FALSE,
  check.mono = TRUE,
  stereo2mono = TRUE,
  overwrite = FALSE,
  freq = 44100,
  round.to = 4
)
}
\arguments{
\item{x}{An Extended data frame to be tagged with media information. See details.}

\item{groupBy}{A variable to group the summary measures. The argument must be a character vector. Default: \code{groupBy = 'wav_path'}.}
}
\description{
Returns summary measures of voice::extract_features
}
\details{
Except by \code{groupBy} and \code{wavPath}, all the parameters are shared with \code{voice::extract_features}.
}
\examples{
library(voice)

# get path to audio file
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
pattern <- glob2rx('*.wav'), full.names = TRUE)

# creating Extended synthetic data
E <- tibble(subject_id = c(1,1,1,2,2,2,3,3,3),
wav_path = path2wav)

# minimal usage
feat_summary(E)

# canonical data
feat_summary(E, 'subject_id')
}
