% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memory}
\alias{memory}
\title{Gives the amount of memory RAM free, total and percentage = free/total.}
\usage{
memory()
}
\description{
Gives the amount of memory RAM free, total and percentage = free/total.
}
\examples{
library(voice)
memory()
}
