% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memory}
\alias{memory}
\title{Gives the amount of memory free, total and percentage.}
\usage{
memory()
}
\arguments{
\item{\code{unit}}{A \code{char} indicating \code{bytes}, \code{kbytes}, \code{mbytes} and \code{gbytes}.}
}
\description{
Gives the amount of memory free, total and percentage.
}
\examples{
library(voice)
memory()
}
