% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{notes}
\alias{notes}
\title{Returns a vector of notes for equal-tempered scale, A4 = 440 Hz.}
\usage{
notes(x, method = "spn", moving.average = FALSE, k = 11)
}
\arguments{
\item{x}{A vector of frequencies in Hz.}

\item{method}{Method of specifying musical pitch. (default: \code{spn}, i.e., Scientific Pitch Notation).}

\item{moving.average}{Logical. Must apply moving average?}

\item{k}{Integer width of the rolling window used if moving.average is TRUE.}
}
\value{
A vector containing the notes for equal-tempered scale, A4 = 440 Hz. When `method = 'spn'` the vector is of class 'ordered factor'. When `method = 'octave'` the vector is of class 'factor'.  When `method = 'midi'` the vector is of class 'integer'.
}
\description{
Returns a vector of notes for equal-tempered scale, A4 = 440 Hz.
}
\details{
The symbol '#' is being used to represent a sharp note, the higher
in pitch by one semitone on Scientific Pitch Notation (SPN)
}
\examples{
library(voice)
notes(c(220,440,880))
notes(c(220,440,880), method = 'octave')
notes(c(220,440,880), method = 'midi')
}
\references{
\url{https://pages.mtu.edu/~suits/notefreqs.html}
}
\seealso{
\code{notes_freq}
}
