% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poetry.R
\name{poetry}
\alias{poetry}
\title{Poetry. The best words in their best order.}
\usage{
poetry(fromWav, to, pycall = "~/miniconda3/envs/pyvoice38/bin/python3.8")
}
\arguments{
\item{fromWav}{A directory/folder containing WAV files.}

\item{to}{A directory/folder to write RTTM files.}

\item{pycall}{Python call.}
}
\description{
Diarization from WAV audios using Python's pyannote-audio library.
}
\examples{
library(voice)
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
                                   pattern <- glob2rx('*.wav'), full.names = TRUE)
dir.create(rttm <- paste0(dirname(path2wav)[1], '/rttm/'))
wsw(from = dirname(path2wav)[1], to = rttm)
}
