% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spoken_time.R
\name{spoken_time}
\alias{spoken_time}
\title{Returns the spoken time of audio files}
\usage{
spoken_time(
  x,
  get.id = FALSE,
  recursive = FALSE,
  pycall = "/home/linuxbrew/.linuxbrew/bin/python3.9"
)
}
\arguments{
\item{x}{Either a WAV file or a directory containing WAV files.}

\item{get.id}{Logical. Should the ID must be extracted from file name? Default: \code{FALSE}.}

\item{recursive}{Logical. Should the listing recursively into directories? (default: \code{FALSE}) Used by \code{base::list.files}.}

\item{pycall}{Python call.}
}
\description{
Returns the spoken time of audio files
}
