% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Tag a data frame with media information.}
\usage{
tag(
  x,
  groupBy = "wav_path",
  wavPath = unique(x$wav_path),
  tags = c("feat_summary"),
  filesRange = NULL,
  features = "f0",
  gender = "u",
  windowShift = 5,
  numFormants = 8,
  numcep = 12,
  dcttype = c("t2", "t1", "t3", "t4"),
  fbtype = c("mel", "htkmel", "fcmel", "bark"),
  resolution = 40,
  usecmp = FALSE,
  mc.cores = 1,
  full.names = TRUE,
  recursive = FALSE,
  check.mono = TRUE,
  stereo2mono = TRUE,
  overwrite = FALSE,
  freq = 44100,
  round.to = 4
)
}
\arguments{
\item{x}{An Extended data frame to be tagged with media information. See details.}

\item{groupBy}{A variable to group the summary measures. Default: \code{groupBy = 1}, the first column.}

\item{tags}{Tags to be added to \code{x}. Default: \code{'feat_summary'}. See details.}

\item{mediaDir}{Directory containing media files. Currently accepts audio files.}

\item{subj.id}{Column containing the subject ID to be searched in mediaDir file names. Default: \code{NULL}, i.e., uses the first column.}

\item{media.id}{Column containing the media ID. Default: \code{NULL}, i.e., uses the first column.}

\item{subj.id.simplify}{Logical. Should subject id must be simplified? Default: \code{FALSE}.}

\item{...}{See \code{?voice::extract_features}.}
}
\description{
Tag a data frame with media information.
}
\details{
Except by \code{groupBy} and \code{wavPath}, all the parameters are shared with \code{voice::extract_features} and \code{voice::feat_summary}.
}
\examples{
library(voice)

# get path to audio file
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
pattern <- glob2rx('*.wav'), full.names = TRUE)

# Extended dataset
E <- dplyr::tibble(subject_id = c(1,1,1,2,2,2,3,3,3),
wav_path = path2wav)

tag(E)
tag(E, 'subject_id')
}
\references{
Zabala, F.J. (2022) to appear in...
}
