% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Tag a data frame with media information.}
\usage{
tag(
  x,
  mediaDir,
  tags = c("feat_summary"),
  subj.id = NULL,
  media.id = NULL,
  subj.id.simplify = FALSE,
  filesRange = NULL,
  features = c("f0"),
  gender = "u",
  windowShift = 5,
  numFormants = 8,
  numcep = 12,
  dcttype = c("t2", "t1", "t3", "t4"),
  fbtype = c("mel", "htkmel", "fcmel", "bark"),
  resolution = 40,
  usecmp = FALSE,
  mc.cores = 1,
  full.names = TRUE,
  recursive = FALSE,
  check.mono = TRUE,
  stereo2mono = TRUE,
  overwrite = FALSE,
  freq = 44100,
  round.to = 4
)
}
\arguments{
\item{x}{An Extended data frame to be tagged with media information. See details.}

\item{mediaDir}{Directory containing media files. Currently accepts audio files.}

\item{tags}{Tags to be added to \code{x}. Default: \code{'feat_summary'}. See details.}

\item{subj.id}{Column containing the subject ID to be searched in mediaDir file names. Default: \code{NULL}, i.e., uses the first column.}

\item{media.id}{Column containing the media ID. Default: \code{NULL}, i.e., uses the first column.}

\item{subj.id.simplify}{Logical. Should subject id must be simplified? Default: \code{FALSE}.}

\item{...}{See \code{?voice::extract_features}.}
}
\description{
Tag a data frame with media information.
}
\details{
Parameter \code{tags} admits c('feat_summary', 'audio_time'). An Extended data frame E must contain subject and media data. See references.
}
\references{
Zabala (2022) to appear in...
}
