% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_list.R
\name{write_list}
\alias{write_list}
\title{Writes a list.}
\usage{
write_list(lista, caminho)
}
\arguments{
\item{\code{lista}}{A list.}

\item{\code{caminho}}{The directory.}
}
\description{
Writes a list.
}
