% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_list.R
\name{write_list}
\alias{write_list}
\title{Writes a list to a path.}
\usage{
write_list(x, path)
}
\arguments{
\item{x}{A list.}

\item{path}{A full path to file.}
}
\value{
A file named `list.txt` in `path`.
}
\description{
Writes a list to a path.
}
\examples{
\dontrun{
library(voice)

pts <- list(x = cars[,1], y = cars[,2])
voice::write_list(pts, paste0(getwd(), '/list.txt'))
}
}
