% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsw.R
\name{wsw}
\alias{wsw}
\title{(Who) Speaks When? Diarization from WAV audios using Python's pyannote-audio library.}
\usage{
wsw(fromWav, to, pycall = "~/miniconda3/envs/py38phdz/bin/python")
}
\arguments{
\item{fromWav}{A directory/folder containing WAV files.}

\item{to}{A directory/folder to write RTTM files.}

\item{pycall}{Python call.}
}
\description{
(Who) Speaks When? Diarization from WAV audios using Python's pyannote-audio library.
}
\examples{
library(voice)
path2wav <- list.files(system.file('extdata', package = 'wrassp'),
                                   pattern <- glob2rx('*.wav'), full.names = TRUE)
dir.create(rttm <- paste0(dirname(path2wav)[1], '/rttm/'))
wsw(from = dirname(path2wav)[1], to = rttm)
}
